requirements.txt
openai
sqlalchemy
pandas
fastapi
uvicorn
python-dotenv
